+++ sandbox/config.js
@@ -15,6 +15,8 @@ export default {
     serviceName: 'Sandbox',
     serviceSubText: 'sandbox.',
     errorContactLink: '#',
-    gaUtmParam: ''
+    gaUtmParam: '',
+    isTextMessageSendingOnQuickReplyDisabled: false
+
   }
 }